fix(docs): make blsag doctest no_std compatible
The previous doctest for bLSAG used a main function returning
'Result<(), Box<dyn std::error::Error>>', which is not compatible
with 'no_std' builds as it relies on the 'std::error::Error' trait.

This change updates the example to remove the 'std' dependency by
changing the main function signature and using '.unwrap()' for the
expected-to-succeed signing operation. This allows the doctest to
pass successfully under the '--no-default-features --features no_std'
configuration.